requirements.txt
google-api-python-client==2.99.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
Pillow==10.0.1
qrcode==7.4.2
tzdata
black
